requirements.txt
# Core Framework
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-jwt-extended==4.6.0

# Database
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
pymongo==4.5.0
redis==5.0.1

# Security & Authentication
werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.4

# API & Documentation
flask-restx==1.2.0
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Monitoring & Logging
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.2

# Data Processing & Analytics
pandas==2.1.1
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Utils
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.0

# Development
pytest==7.4.2
pytest-flask==1.2.1
black==23.7.0
flake8==6.0.0
